1.添加eidea验证类库 2.修复查询条件新建会删除其他查询条件明细的bug

+++ ebase/src/main/java/com/dsdl/eidea/core/service/impl/SearchServiceImpl.java
@@ -111,7 +111,10 @@ public class SearchServiceImpl implements SearchService {
                 notNeedRemovedColumnList.add(searchColumnPo.getId());
             }
         }
-        searchColumnDao.removeSearchColumnIdNotIn(notNeedRemovedColumnList, searchBo.getId());
+        if( searchBo.getId()!=null)
+        {
+            searchColumnDao.removeSearchColumnIdNotIn(notNeedRemovedColumnList, searchBo.getId());
+        }
         searchPo.setCoreSearchColumns(searchColumnPoList);
 
         searchDao.save(searchPo);
